+++ front/src/App.jsx
@@ -1,39 +1,14 @@
 import axios from "axios";
 
 import { useEffect, useState } from "react";
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
-import ButtonAppBar from './components/Appbar';
-import { Gallery } from './components/Gallery';
 import SignUp from './SignUp';
-import useAddModal from './hooks/useAddModal';
 import { AuthProvider } from './context/AuthContext';
 import NoMatch from './routes/NoMatch';
 import { Home } from './pages/Home';
 
 // import { BrowserRouter, Link, Switch, Route } from "react-router-dom";
-
-
-
-function Header(props){
-  const { onAddClick } = props;
-
-  return (
-    <ButtonAppBar onAddClick={onAddClick}/>
-  );
-}
-
-function Clothet(props) {
-  const { urls, onLinkClick, onDeleteClick } = props;
-
-  return (
-    <Gallery
-    urls={urls}
-    onLinkClick={onLinkClick}
-    onDeleteClick={onDeleteClick}
-    />
-  );
-}
 
 function App() {
 	// const GetData = () => {
