+++ front/src/components/AddModal.jsx
@@ -1,21 +1,30 @@
 import useAddModal from "../hooks/useAddModal";
+import { AddButton } from "./AddButton";
 
-export default function AddModal() {
-    const { Modal, closeAddModal } = useAddModal();
+function AddModal() {
+
+    const { AddModal, openAddModal, closeAddModal } = useAddModal();
+
     return (
-        <Modal>
-            <div
-                style={{
-                backgroundColor: 'white',
-                width: '300px',
-                height: '200px',
-                padding: '1em',
-                borderRadius: '15px',
-            }}
-            >
-            <h2>追加してください</h2>
-            <button onClick={closeAddModal}>Close</button>
+        <div>
+            <div>
+                <AddButton onAddClick={openAddModal} />
             </div>
-        </Modal>
+            <AddModal>
+                <div
+                    style={{
+                    backgroundColor: 'white',
+                    width: '300px',
+                    height: '200px',
+                    padding: '1em',
+                    borderRadius: '15px',
+                }}
+                >
+                <h2>追加してください</h2>
+                <button onClick={closeAddModal}>Close</button>
+                </div>
+            </AddModal>
+        </div>
     );
-}
+}    
+export { AddModal };
