+++ front/src/layout/Clothet.jsx
@@ -1,0 +1,16 @@
+import { Gallery} from "../components/Gallery";
+
+
+function Clothet(props) {
+    const { urls, onLinkClick, onDeleteClick } = props;
+
+    return (
+        <Gallery
+        urls={urls}
+        onLinkClick={onLinkClick}
+        onDeleteClick={onDeleteClick}
+        />
+);
+}
+
+export { Clothet };
