+++ front/src/layout/Header.jsx
@@ -1,0 +1,12 @@
+import { ButtonAppBar } from '../components/AppBar';
+
+
+function Header(props){
+    const { onAddClick } = props;
+  
+    return (
+      <ButtonAppBar onAddClick={onAddClick}/>
+    );
+  }
+
+export { Header };
